+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- Add a command `firebase firestore:databases:clone` to clone a Firestore database (#9262).
 - Add JSON format support for Cloud Functions secrets with `--format json` flag and auto-detection from file extensions (#1745)
 - `firebase dataconnect:sdk:generate` will run `init dataconnect:sdk` automatically if no SDKs are configured (#9325).
 - Tighten --only filter resolution for functions deployment to prefer codebase names (#9353)